+++ src/coverlet.collector/Utilities/CoverletConstants.cs
@@ -26,5 +26,7 @@ namespace Coverlet.Collector.Utilities
         public const string SkipAutoProps = "SkipAutoProps";
         public const string DoesNotReturnAttributesElementName = "DoesNotReturnAttribute";
         public const string DeterministicReport = "DeterministicReport";
+        public const string IncludeTargetFrameworkElementName = "IncludeTargetFramework";
+        public const string Framework = "Framework";
     }
 }