+++ test/coverlet.integration.tests/Collectors.cs
@@ -88,7 +88,7 @@ namespace Coverlet.Integration.Tests
         public void TestVsTest_Test_Settings()
         {
             using ClonedTemplateProject clonedTemplateProject = PrepareTemplateProject();
-            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!);
+            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!, reportNameWithFramework: false);
             Assert.True(DotnetCli($"test -c {_buildConfiguration} \"{clonedTemplateProject.ProjectRootPath}\" --collect:\"XPlat Code Coverage\" --settings \"{runSettingsPath}\" --diag:{Path.Combine(clonedTemplateProject.ProjectRootPath, "log.txt")}", out string standardOutput, out string standardError), standardOutput);
             Assert.Contains("Passed!", standardOutput);
             AssertCoverage(clonedTemplateProject);
@@ -99,7 +99,7 @@ namespace Coverlet.Integration.Tests
         public void TestVsTest_VsTest()
         {
             using ClonedTemplateProject clonedTemplateProject = PrepareTemplateProject();
-            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!);
+            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!, reportNameWithFramework: false);
             Assert.True(DotnetCli($"publish -c {_buildConfiguration} {clonedTemplateProject.ProjectRootPath}", out string standardOutput, out string standardError), standardOutput);
             string publishedTestFile = clonedTemplateProject.GetFiles("*" + ClonedTemplateProject.AssemblyName + ".dll").Single(f => f.Contains("publish"));
             Assert.NotNull(publishedTestFile);
@@ -114,7 +114,7 @@ namespace Coverlet.Integration.Tests
         public void TestVsTest_VsTest_Settings()
         {
             using ClonedTemplateProject clonedTemplateProject = PrepareTemplateProject();
-            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!);
+            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!, reportNameWithFramework: false);
             Assert.True(DotnetCli($"publish -c {_buildConfiguration} \"{clonedTemplateProject.ProjectRootPath}\"", out string standardOutput, out string standardError), standardOutput);
             string publishedTestFile = clonedTemplateProject.GetFiles("*" + ClonedTemplateProject.AssemblyName + ".dll").Single(f => f.Contains("publish"));
             Assert.NotNull(publishedTestFile);
@@ -122,6 +122,32 @@ namespace Coverlet.Integration.Tests
             Assert.Contains("Passed!", standardOutput);
             AssertCoverage(clonedTemplateProject);
             AssertCollectorsInjection(clonedTemplateProject);
+        }
+
+        [Fact]
+        public void TestVsTest_Test_SettingsWithFrameworkNameInReport()
+        {
+            using ClonedTemplateProject clonedTemplateProject = PrepareTemplateProject();
+            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!, reportNameWithFramework: true);
+            Assert.True(DotnetCli($"test -c {_buildConfiguration} \"{clonedTemplateProject.ProjectRootPath}\" --collect:\"XPlat Code Coverage\" --settings \"{runSettingsPath}\" --diag:{Path.Combine(clonedTemplateProject.ProjectRootPath, "log.txt")}", out string standardOutput, out string standardError), standardOutput);
+            Assert.Contains("Passed!", standardOutput);
+            AssertCoverage(clonedTemplateProject, filter: "coverage.net5.0.json");
+            AssertCollectorsInjection(clonedTemplateProject);
+        }
+
+        [Fact]
+        public void TestVsTest_Test_SettingsWithFrameworkNameInReportForMultiTargetProject()
+        {
+            using ClonedTemplateProject clonedTemplateProject = PrepareTemplateProject();
+            string runSettingsPath = AddCollectorRunsettingsFile(clonedTemplateProject.ProjectRootPath!, reportNameWithFramework: true);
+            string[] targetFrameworks = new string[] { "net5.0", "netcoreapp3.1" };
+            UpdateProjectTargetFramework(clonedTemplateProject, targetFrameworks);
+
+            Assert.True(DotnetCli($"test -c {_buildConfiguration} \"{clonedTemplateProject.ProjectRootPath}\" --collect:\"XPlat Code Coverage\" --settings \"{runSettingsPath}\" --diag:{Path.Combine(clonedTemplateProject.ProjectRootPath, "log.txt")}", out string standardOutput, out string standardError), standardOutput);
+
+            Assert.Contains("Passed!", standardOutput);
+            AssertCoverage(clonedTemplateProject, filter: "coverage.net5.0.json");
+            AssertCoverage(clonedTemplateProject, filter: "coverage.netcoreapp3.1.json");
         }
     }
 }