+++ Documentation/Changelog.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -ExcludeFromCodeCoverage attribute on local functions ignores lambda expression [#1302](https://github.com/coverlet-coverage/coverlet/issues/1302)
 
 ### Improvements  
--Added new parameter to control automatic assembly exclusion and changed the default behavior [#1164](https://github.com/coverlet-coverage/coverlet/issues/1164)
+-Added new parameter to control automatic assembly exclusion and changed the default behavior [#1164](https://github.com/coverlet-coverage/coverlet/issues/1164) 
 
 ## Release date 2022-02-06
 ### Packages